pyproject.toml
[project]
name = "text2cypher"
version = "0.1.0"
description = "Clinical note generation with T5 and PEFT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch>=2.0",
  "pytorch-lightning>=2.0",
  "transformers>=4.38",
  "peft>=0.8.0",
  "datasets",
  "hydra-core>=1.3",
  "loguru",
  "wandb",
  "omegaconf>=2.3",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["text2cypher"]

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "pytest",
  "mypy",
  "ipykernel"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I"]
ignore = ["E501"]  # Ignore line too long

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
  "tests",
]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
